Collect external option interfaces in src/options.ts: SchemaOptions, CreateNodeOptions, ToJSOptions
